[priority sampling] splitting distributed_headers and http_propagator
